# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BE API — WP-02 Identidad y sesiones · WP-03 Vínculo, consentimiento y PDP",
    "version": "0.3.0",
    "description": "Generado desde @be/domain (contratos.ts y contratos-vinculo.ts). No editar a mano."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "sesion": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Sesión revocable verificada en cada request (08 §26)."
      }
    },
    "schemas": {
      "ErrorEnvelope": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
...
    "/relationship-requests": {
      "post": {
        "operationId": "API-REL-01",
        "summary": "Crear una solicitud de vínculo por alcance y finalidad. No concede acceso. Una equivalente pendiente se devuelve con 200 deduplicado (REG-06-44).",
        "description": "Fuente: 09v8:1108-1192 · DEUDA_LEGAJO DL-035/DL-037. AuthN: SESSION.",
        "security": [
          {
            "sesion": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{8,128}$"
            }
          },
          {
            "name": "X-BE-Surface",
            "in": "header",
            "required": false,
            "description": "Superficie declarada (procedencia). Nunca autoriza (DEUDA_LEGAJO DL-022).",
            "schema": {
              "type": "string",
              "enum": [
                "WEB",
                "APK"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "target": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "PROFESSIONAL",
                          "ADVISEE"
                        ]
                      },
                      "identityId": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": [
                      "type",
                      "identityId"
                    ],
                    "additionalProperties": false
                  },
                  "scope": {
                    "type": "object",
                    "properties": {
                      "code": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "code"
                    ],
                    "additionalProperties": false
                  },
                  "purpose": {
                    "type": "string"
                  }
                },
                "required": [
                  "target",
                  "scope",
                  "purpose"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "relationshipRequestId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "deduplicated": {
                          "type": "boolean",
                          "const": true
                        }
                      },
                      "required": [
                        "relationshipRequestId",
                        "deduplicated"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "Éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "relationshipRequestId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "version": {
                          "type": "string",
                          "pattern": "^v[1-9]\\d*$"
                        },
                        "professional": {
                          "type": "object",
                          "properties": {
                            "identityId": {
                              "type": "string",
                              "minLength": 1
                            },
                            "displayName": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "identityId",
                            "displayName"
                          ],
                          "additionalProperties": false
                        },
                        "advisee": {
                          "type": "object",
                          "properties": {
                            "identityId": {
                              "type": "string",
                              "minLength": 1
                            },
                            "displayName": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "identityId",
                            "displayName"
                          ],
                          "additionalProperties": false
                        },
                        "scope": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "NUTRICION",
                                "ENTRENAMIENTO",
                                "ANTROPOMETRIA"
                              ]
                            },
                            "label": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "code",
                            "label"
                          ],
                          "additionalProperties": false
                        },
                        "purpose": {
                          "type": "string",
                          "enum": [
                            "ACOMPANAMIENTO_NUTRICIONAL",
                            "PLANIFICACION_DEL_ENTRENAMIENTO",
                            "EVALUACION_ANTROPOMETRICA"
                          ]
                        },
                        "state": {
                          "type": "string",
                          "enum": [
                            "PENDIENTE",
                            "ACEPTADA",
                            "RECHAZADA",
                            "CADUCADA",
                            "INVALIDADA"
                          ]
                        },
                        "initiatedBy": {
                          "type": "string",
                          "enum": [
                            "PROFESSIONAL",
                            "ADVISEE"
                          ]
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                        },
                        "expiresAt": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date-time",
                              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "relationshipRequestId",
                        "version",
                        "professional",
                        "advisee",
                        "scope",
                        "purpose",
                        "state",
                        "initiatedBy",
                        "createdAt",
                        "expiresAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "ErrorEnvelope: INVALID_REQUEST | UNKNOWN_FIELD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "ErrorEnvelope: AUTHENTICATION_REQUIRED | SESSION_INVALID | SESSION_EXPIRED | SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "ErrorEnvelope: RESOURCE_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "ErrorEnvelope: IDEMPOTENCY_KEY_REUSED | RESOURCE_CONFLICT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "ErrorEnvelope: PURPOSE_REQUIRED | SCOPE_NOT_AVAILABLE | COUNTERPART_NOT_ELIGIBLE | VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "ErrorEnvelope: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "ErrorEnvelope: DB_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/me/relationship-requests": {
      "get": {
        "operationId": "API-REL-02",
        "summary": "Solicitudes propias, enviadas y recibidas. No incluye información sanitaria.",
        "description": "Fuente: 09v8:1196-1224. AuthN: SESSION.",
        "security": [
          {
            "sesion": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Tamaño de página (1 a 50; 20 por defecto).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "Cursor opaco de la página siguiente.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "description": "Estado de la solicitud (06 §7.3.1).",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "ACEPTADA",
                "RECHAZADA",
                "CADUCADA",
                "INVALIDADA"
              ]
            }
          },
          {
            "name": "X-BE-Surface",
            "in": "header",
            "required": false,
            "description": "Superficie declarada (procedencia). Nunca autoriza (DEUDA_LEGAJO DL-022).",
            "schema": {
              "type": "string",
              "enum": [
                "WEB",
                "APK"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Éxito",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "relationshipRequestId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "version": {
                            "type": "string",
                            "pattern": "^v[1-9]\\d*$"
                          },
                          "professional": {
                            "type": "object",
                            "properties": {
                              "identityId": {
                                "type": "string",
                                "minLength": 1
                              },
                              "displayName": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "identityId",
                              "displayName"
                            ],
                            "additionalProperties": false
                          },
                          "advisee": {
                            "type": "object",
                            "properties": {
                              "identityId": {
                                "type": "string",
                                "minLength": 1
                              },
                              "displayName": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "identityId",
                              "displayName"
                            ],
                            "additionalProperties": false
                          },
                          "scope": {
                            "type": "object",
                            "properties": {
                              "code": {
                                "type": "string",
                                "enum": [
                                  "NUTRICION",
                                  "ENTRENAMIENTO",
                                  "ANTROPOMETRIA"
                                ]
                              },
                              "label": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "code",
                              "label"
                            ],
                            "additionalProperties": false
                          },
                          "purpose": {
                            "type": "string",
                            "enum": [
                              "ACOMPANAMIENTO_NUTRICIONAL",
                              "PLANIFICACION_DEL_ENTRENAMIENTO",
                              "EVALUACION_ANTROPOMETRICA"
                            ]
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "PENDIENTE",
                              "ACEPTADA",
                              "RECHAZADA",
                              "CADUCADA",
                              "INVALIDADA"
                            ]
                          },
                          "initiatedBy": {
                            "type": "string",
                            "enum": [
                              "PROFESSIONAL",
                              "ADVISEE"
                            ]
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                          },
                          "expiresAt": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "date-time",
                                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        },
                        "required": [
                          "relationshipRequestId",
                          "version",
                          "professional",
                          "advisee",
                          "scope",
                          "purpose",
                          "state",
                          "initiatedBy",
                          "createdAt",
                          "expiresAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "page": {
                      "type": "object",
                      "properties": {
                        "limit": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "maximum": 9007199254740991
                        },
                        "nextCursor": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "limit",
                        "nextCursor",
                        "hasMore"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "page"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "ErrorEnvelope: INVALID_REQUEST | INVALID_CURSOR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "ErrorEnvelope: AUTHENTICATION_REQUIRED | SESSION_INVALID | SESSION_EXPIRED | SESSION_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "ErrorEnvelope: INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "ErrorEnvelope: DB_UNAVAILABLE",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}